refactor: separate folder for mocks
