+++ .gitignore
@@ -81,3 +81,4 @@ typings/
 .fusebox/
 
 # End of https://www.gitignore.io/api/node
+.tool-versions
