+++ src/aws-sdk/aws-sdk-mock.js
