+++ src/aws-sdk/cognito-isp.js
