+++ src/index.js
@@ -1,4 +1,4 @@
-import awsSdkMock from './aws-sdk-mock';
+import awsSdkMock from './aws-sdk/aws-sdk-mock';
 
 export {
   awsSdkMock
